# Improve CDATA and comment parse performance