--- a/lib/rexml/parsers/baseparser.rb
+++ b/lib/rexml/parsers/baseparser.rb
@@ -449,9 +449,7 @@ module REXML
               end
               return [ :end_element, last_tag ]
             elsif @source.match?("!", true)
-              md = @source.match(/([^>]*>)/um)
               #STDERR.puts "SOURCE BUFFER = #{source.buffer}, #{source.buffer.size}"
-              raise REXML::ParseException.new("Malformed node", @source) unless md
               if @source.match?("--", true)
                 return [ :comment, process_comment ]
               elsif @source.match?("[CDATA[", true)
@@ -461,9 +459,9 @@ module REXML
                 else
                   raise REXML::ParseException.new("Malformed CDATA: Missing end ']]>'", @source)
                 end
+              else
+                raise REXML::ParseException.new("Malformed node: Started with '<!' but not a comment nor CDATA", @source)
               end
-              raise REXML::ParseException.new( "Declarations can only occur "+
-                "in the doctype declaration.", @source)
             elsif @source.match?("?", true)
               return process_instruction
             else